+++ ex-treemap/test_ctags.py
@@ -1,7 +1,7 @@
-import make_index
+import make_ctags
 
 
 def test_compile():
-    df = make_index.compile(None, paths=['simple.py'])
+    df = make_ctags.compile(None, paths=['simple.py'])
     assert len(df) == 1
     assert df.columns.tolist() == ['path', 'name', 'ctags_raw', 'linecount']
